[BUGFIX] Use lowercase for generalutility class name
Since the extension supports TYPO3 4.5+ and the extension structure
and class names don't support implicit class autoloading, the classes
are registered in ext_autoload.php. In ext_autoload.php, the class
names need to be in lowercase (this limitation was removed in TYPO3
6.0). As a consequence, "tx_varnish_GeneralUtility" is named
"tx_varnish_generalutility" in the ext_autoload.php file. This is
why in some cases TYPO3 fails to call the class, leading to a fatal
error.

To remain backwards compatibility, but the class and all references to
it should be written in lowercase.