[BUGFIX] Use t3lib_extMgm::addTypoScript() for adding TypoScript setup
When t3lib_extMgm::addTypoScriptSetup() is called, it may happen that the
rowSum has a different value after $TSFE->tmpl->runThroughTemplates() and
after $TSFE->tmpl->generateConfig(). The result is that the page is cached
but not found due to looking for a wrong hash.

This seems to be a bug in t3lib_extMgm::addTypoScriptSetup() which needs
more investigation. It also seems like the function is rarely used so it
may not have been noticed yet.
For now, the workaround is to call t3lib_extMgm::addTypoScript() which
works fine.